Eliminated some more warnings when compiling with gcc/mingw

+++ src/pdffontparsertype1.cpp
@@ -1206,7 +1206,7 @@ wxPdfFontParserType1::ReadPFM(wxInputStream& pfmFile)
   int italicAngle = 0;
   if (hdr.italic != 0 || fontNameLower.Find(wxT("italic")) != wxNOT_FOUND)
   {
-    italicAngle = ext.slant / 10.0;
+    italicAngle = int(ext.slant / 10);
     // -12.00 - this is a typical value
   }
   fd.SetItalicAngle(italicAngle);
